+++ index.js
@@ -1,9 +1,9 @@
 if (Number(process.version.slice(1).split('.')[0]) < 8) throw new RangeError('Node 8.0.0 or higher is required. Update Node on your system.');
 
 const Discord = require('discord.js'); 
-const sentryconfig = require('./sentry.json');
-const Raven = require('raven');
-Raven.config(sentryconfig.link).install();
+//const sentryconfig = require('./sentry.json');
+//const Raven = require('raven');
+//Raven.config(sentryconfig.link).install();
 const { promisify } = require('util');
 const readdir = promisify(require('fs').readdir);
 const Enmap = require('enmap');
@@ -11,7 +11,10 @@ const EnmapLevel = require('enmap-sqlite'); //eslint-disable-line no-unused-vars
 
 const client = new Discord.Client({
   fetchAllMembers: true,
-  disableEveryone: true}); 
+  disableEveryone: true,
+  //shardId: process.argv[1],
+  //shardCount: process.argv0[2],
+}); 
 
 
 client.config = require('./config.js');
@@ -19,9 +22,9 @@ client.config = require('./config.js');
 // client.config.prefix contains the message prefix
 //k
 
-client.logger = require('./modules/Logger');
+client.logger = require('./modules/Logger.ts');
 
-require('./modules/functions.js')(client);
+require('./modules/functions.ts')(client);
 
 client.commands = new Enmap();
 client.aliases = new Enmap();
@@ -77,7 +80,7 @@ client.userBio = new Enmap({
   fetchAll: true
 });
 
-const init = async () => {
+const init =  async () => {
 
   const { join } = require('path');
   const commands = await readdir(join(__dirname, './commands/'));
